Adding performance tests for PRC functionality.
