--- DtronixMessageQueue.Tests.Performance/Run.bat
@@ -1,2 +1,0 @@
-DMQPerf.exe single-process 1 100000 4 50 10
-pause
