+++ DtronixMessageQueue.Tests.Performance/SimpleRpcSession.cs
@@ -1,0 +1,8 @@
+﻿using System;
+using DtronixMessageQueue.Rpc;
+
+namespace DtronixMessageQueue.Tests.Performance {
+
+	public class SimpleRpcSession : RpcSession<SimpleRpcSession> {
+	}
+}
