+++ DtronixMessageQueue.Tests.Performance/run-mq.bat
@@ -1,0 +1,2 @@
+DMQPerf.exe mq single-process 1 100000 4 50 10
+pause
