+++ DtronixMessageQueue.Tests.Performance/run-rpc.bat
@@ -1,0 +1,2 @@
+DMQPerf.exe rpc single-process 1 100000 4 50 10
+pause
